# Optimize devices lua and python lib