@fnuecke 2022-07-17
LGTM, thanks! I'm worried that the issues with random broken bytes is a bug in sedna -- there's also an as yet unresolved issue with TCC randomly segfaulting in sedna (while it doesn't in Qemu using the same rootfs and kernel). I've not yet been successful in figuring out why that happens. I'll give this a test-spin soon.