@ToMe25 2022-07-23
They could be removed, I just thought not having breaking API changes would be better.

I'm not at home for the next week, so if you want them removed soon, please feel free to remove them yourself :)
Otherwise just let me know if they should be removed and I will remove them.