@fnuecke 2022-07-24
Since they're "private" I wouldn't really consider them part of the API, I'll remove them then :)